+++ packages/api-proxy/src/platform/api/create-selector-query/index.android.js
@@ -1,0 +1,1 @@
+export * from './index.ios'
