+++ packages/api-proxy/src/platform/api/create-selector-query/index.ios.js
@@ -1,0 +1,9 @@
+import SelectorQuery from './rnSelectQuery'
+
+function createSelectorQuery () {
+  return new SelectorQuery()
+}
+
+export {
+  createSelectorQuery
+}
